backend/requirements.txt
flask
flask-cors
spacy
nltk
textstat
requests
python-dotenv
